+++ tests/e2e/greenfield/test_object.py
@@ -425,6 +425,69 @@ async def test_create_object_and_set_tag():
 @pytest.mark.requires_config
 @pytest.mark.tx
 @pytest.mark.slow
+async def test_create_object_with_tag():
+    config = get_account_configuration()
+    key_manager = KeyManager(private_key=config.private_key)
+    async with GreenfieldClient(network_configuration=network_configuration, key_manager=key_manager) as client:
+        bucket_name = "".join(random.choices(string.ascii_lowercase + string.digits, k=random.randint(5, 11)))
+        object_name = "".join(random.choices(string.ascii_lowercase + string.digits, k=random.randint(5, 11)))
+
+        await client.async_init()
+        sp = (await client.blockchain_client.get_active_sps())[0]
+        tx_hash = await client.bucket.create_bucket(
+            bucket_name,
+            primary_sp_address=sp["operator_address"],
+            opts=CreateBucketOptions(charged_read_quota=100, visibility=VisibilityType.VISIBILITY_TYPE_PRIVATE),
+        )
+        assert tx_hash
+        assert len(tx_hash) == 64
+        assert isinstance(tx_hash, str)
+        await client.basic.wait_for_tx(hash=tx_hash)
+
+        tags = ResourceTags(tags=[ResourceTagsTag(key="tag1", value="first_tag")])
+        tx_hash = await client.object.create_object(
+            bucket_name,
+            object_name,
+            CONTENT,
+            opts=CreateObjectOptions(tags=tags),
+        )
+        assert tx_hash
+        assert len(tx_hash) == 64
+        assert isinstance(tx_hash, str)
+        await client.basic.wait_for_tx(hash=tx_hash)
+
+        put_object = await client.object.put_object(
+            bucket_name,
+            object_name,
+            CONTENT.getbuffer().nbytes,
+            CONTENT.getvalue(),
+            opts=PutObjectOptions(),
+        )
+        assert put_object == "Object added successfully"
+
+        await asyncio.sleep(8)
+
+        head_object = await client.object.get_object_head(bucket_name, object_name)
+        assert head_object
+        assert head_object.tags == tags
+
+        tx_hash = await client.object.delete_object(
+            bucket_name,
+            object_name,
+        )
+        assert tx_hash
+        assert len(tx_hash) == 64
+        assert isinstance(tx_hash, str)
+        await client.basic.wait_for_tx(hash=tx_hash)
+
+        tx_hash = await client.bucket.delete_bucket(bucket_name)
+        assert tx_hash
+        await client.basic.wait_for_tx(hash=tx_hash)
+
+
+@pytest.mark.requires_config
+@pytest.mark.tx
+@pytest.mark.slow
 @pytest.mark.go_library
 async def test_cancel_creation_object():
     config = get_account_configuration()